Organized Folder Structure, and Added Redux-Logger
