+++ package.json
@@ -18,6 +18,7 @@
     "redux",
     "react-redux",
     "react-router",
+    "redux-logger",
     "react-router-dom",
     "typescript",
     "prettier"
@@ -64,6 +65,7 @@
     "react-scripts": "4.0.0",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
+    "redux-logger": "^3.0.6",
     "redux-mock-store": "^1.5.4",
     "typescript": "~4.0.3"
   },
@@ -110,6 +112,7 @@
     "@semantic-release/changelog": "5.0.1",
     "@semantic-release/git": "9.0.0",
     "@semantic-release/github": "7.2.0",
+    "@types/redux-logger": "^3.0.8",
     "codecov": "3.8.1",
     "commitizen": "4.2.2",
     "cypress": "5.4.0",
