+++ src/redux/counter/index.ts
