+++ template/src/components/counter/Counter.spec.tsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux'
 import { mount } from 'enzyme'
 import configureStore from 'redux-mock-store'
 
-import { actionTypes } from '../../features/counter'
+import { actionTypes } from '../../redux/counter'
 import Counter from './Counter'
 
 describe('Counter', () => {
