+++ template/src/redux/counter/actionTypes.ts
