+++ template/src/redux/counter/counterReducer.spec.ts
