+++ template/src/redux/counter/counterReducer.ts
