+++ template/src/redux/counter/index.ts
